// web/src/lib/api/generated/models/internal-orders-order-list-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * POS Kasir API
 * POS Kasir API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { InternalOrdersOrderItemResponse } from './internal-orders-order-item-response';
// May contain unused imports in some cases
// @ts-ignore
import type { POSKasirInternalOrdersRepositoryOrderStatus } from './poskasir-internal-orders-repository-order-status';
// May contain unused imports in some cases
// @ts-ignore
import type { POSKasirInternalOrdersRepositoryOrderType } from './poskasir-internal-orders-repository-order-type';

export interface InternalOrdersOrderListResponse {
    'created_at'?: string;
    'id'?: string;
    'is_paid'?: boolean;
    'items'?: Array<InternalOrdersOrderItemResponse>;
    'net_total'?: number;
    'queue_number'?: string;
    'status'?: POSKasirInternalOrdersRepositoryOrderStatus;
    'type'?: POSKasirInternalOrdersRepositoryOrderType;
    'user_id'?: string;
}